remove legacy statistics service test
