+++ tests/test_core/test_single_cloud_service.py
@@ -1,14 +1,14 @@
-"""Tests for StatisticsService.calc_single_cloud_stats output ordering."""
+"""Tests for ``single_cloud_service.calc_single_cloud_stats`` output ordering."""
 
 import sys
 from pathlib import Path
-from typing import List, Dict
+from typing import Dict, List
 
 import pandas as pd
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
-from m3c2.m3c2_core.statistics import service
+from m3c2.statistics import single_cloud_service as service
 
 
 def test_calc_single_cloud_stats_metadata_first(monkeypatch):
@@ -29,10 +29,10 @@ def test_calc_single_cloud_stats_metadata_first(monkeypatch):
     def fake_write_cloud_stats(df, out_path, sheet_name, output_format):
         captured_dfs.append(df)
 
-    monkeypatch.setattr(service, "calc_single_cloud_stats", fake_calc_single_cloud_stats)
+    monkeypatch.setattr(service, "_calc_single_cloud_stats", fake_calc_single_cloud_stats)
     monkeypatch.setattr(service, "write_cloud_stats", fake_write_cloud_stats)
 
-    service.StatisticsService.calc_single_cloud_stats(
+    service.calc_single_cloud_stats(
         folder_ids=["fid"],
         filename_singlecloud="file",
         singlecloud=None,
